+++ docs/CHANGELOG.md
@@ -1,9 +1,17 @@
 ## [unreleased] - 2021-07-09
  ### Added
 
+ - Add `min_consensus_read` argument to `celescope.tools.consensus`. If 
+    1. the percentage of the most common residue type > threshold;
+    2. most common residue reads >= min_consensus_read;
+    
+then we will add that residue type, otherwise an ambiguous character will be added.
+
  ### Changed
 
  ### Fixed
+
+ - Fix an error when there is no `Undetermined` cells in `celescope tag count_tag`
 
  ### Removed