+++ docs/CHANGELOG.md
@@ -7,11 +7,13 @@
     
 then we will add that residue type, otherwise an ambiguous character will be added.
 
+- Use otsu method to calculate `min_support_read` for `capture_virus`.
+
  ### Changed
 
  ### Fixed
 
- - Fix an error when there is no `Undetermined` cells in `celescope tag count_tag`
+ - Fix an error when there is no `Undetermined` cells in `celescope tag count_tag`.
 
  ### Removed